backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distlib==0.3.9
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.15.2
Faker==37.0.0
filelock==3.16.1
idna==3.10
nulltype==2.3.1
plaid-python==27.0.0
platformdirs==4.3.6
psycopg2==2.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
virtualenv==20.27.0
gunicorn>=20.1.0
